Aggregates amount donated and calculates hours sponsored.

+++ home/views.py
@@ -18,7 +18,7 @@ def home(request):
     posts = Post.objects.filter(is_published=True)[:5]
     donations = Donation.objects.all()
     donated_amount = donations.aggregate(Sum('amount'))['amount__sum']
-    hours = floor(donated_amount / 10)
+    hours = int(floor(donated_amount / 10))
     return render_to_response('index.html',
         {'posts': posts, 'donations': donations, 'form': form, 'donated_amount': donated_amount, 'hours': hours},
         context_instance=RequestContext(request))
